torch.float32->torch.float64

+++ swimmer_chemotaxis_data.py
@@ -29,8 +29,8 @@ def generate_sample(env, model):
         if done:
             break
     env.close()
-    x = torch.tensor(env.stats['concentration'], dtype=torch.float32).unsqueeze(-1)
-    y = torch.tensor(y, dtype=torch.float32)
+    x = torch.tensor(env.stats['concentration'], dtype=torch.float64).unsqueeze(-1)
+    y = torch.tensor(y, dtype=torch.float64)
     return x, y
